try fix cast error

+++ mock-nm.cpp
@@ -5,7 +5,7 @@
 
 static void setActiveConnectionState(gpointer state)
 {
-	NMActiveConnectionState activeConnectionState = (NMActiveConnectionState)state;
+	NMActiveConnectionState activeConnectionState = (NMActiveConnectionState)(int)state;
 	NMActiveConnection* activeConnection = configMockNM.getActiveConnection();
 	if (nm_active_connection_get_state(activeConnection) != activeConnectionState)
 	{
